Refactor standalone Windows Nsight discovery
